+++ CHANGELOG.md
@@ -13,3 +13,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Creepers not using the spawn weight/group size from config
 ### Added 
 - CreeperTypes json now allows biomesTags list for handling spawns
+- Config option to nerf drops when automated
+- Config option to nerf drop amounts when automated
+- Config option to set the nerf-percentage when automated